+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Changed - Default constructed T() added to optVec&lt;T> before parse action
 - Added - cli.badRange() helper
 - Added - opt.siUnits(), opt.timeUnits(), and opt.anyUnits()
+- Added - Allow limiting the number of values for vector options
 
 ## dimcli 4.1.0 (2018-12-05)
 - Fixed - Order of keys in help text of multi-keyed options non-deterministic